MBX-3635: Fix requests
